recommendation-engine/requirements.txt
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
exceptiongroup==1.3.1
Flask==3.1.3
flask-cors==6.0.2
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.1
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.20.0
pytest==9.0.3
python-dotenv==1.2.2
requests==2.33.1
tomli==2.4.1
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.8
